Basic Stylization and Edited Forms
I added the style and a few other css files for the basic overall style for the project, ontop of that I edited some of the forms as well sa their formats as well. Lastly wrote a basic README for us to have displayed. Hopefully I didn't mess anything up but I tested everything on a local version so it should be okay.